Factor code into new entity base class
